HopSlot-FullStack/django-api/predict_symptoms/views.py
from django.shortcuts import render
from rest_framework.response import Response
from rest_framework.decorators import api_view
from rest_framework.views import APIView


# Create your views here.
class PredictSymptomsAPIView(APIView):
    def get(self, request, *args, **kwargs):
        '''
        List all the todo items for given requested user
        '''
        return Response({"test": "asasd"})
