HopSlot-FullStack/django-api/requirements.txt
Django
djangorestframework
scikit-learn
joblib
django-environ